+++ launchdarkly/webhooks_helper.go
@@ -22,8 +22,16 @@ func baseWebhookSchema() map[string]*schema.Schema {
 			Optional:    true,
 			Description: "A human-readable name for your webhook",
 		},
-		POLICY_STATEMENTS: policyStatementsSchema(),
-		TAGS:              tagsSchema(),
+		POLICY_STATEMENTS: policyStatementsSchema(
+			policyStatementSchemaOptions{
+				deprecated:    "'policy_statements' is deprecated in favor of 'statements'",
+				conflictsWith: []string{STATEMENTS},
+			},
+		),
+		STATEMENTS: policyStatementsSchema(policyStatementSchemaOptions{
+			conflictsWith: []string{POLICY_STATEMENTS},
+		}),
+		TAGS: tagsSchema(),
 	}
 }
 
@@ -69,9 +77,29 @@ func webhookRead(d *schema.ResourceData, meta interface{}, isDataSource bool) er
 	}
 
 	_ = d.Set(NAME, webhook.Name)
-	err = d.Set(POLICY_STATEMENTS, statements)
-	if err != nil {
-		return fmt.Errorf("failed to set policy_statements on webhook with id %q: %v", webhookID, err)
+
+	// // "policy_statements" is deprecated in favor of "statements". For data sources, set both, for resources only set the one being used.
+	if isDataSource {
+		err = d.Set(POLICY_STATEMENTS, statements)
+		if err != nil {
+			return fmt.Errorf("failed to set policy_statements on webhook with id %q: %v", webhookID, err)
+		}
+		err = d.Set(STATEMENTS, statements)
+		if err != nil {
+			return fmt.Errorf("failed to set statements on webhook with id %q: %v", webhookID, err)
+		}
+	} else {
+		if _, ok := d.GetOk(POLICY_STATEMENTS); ok {
+			err = d.Set(POLICY_STATEMENTS, statements)
+			if err != nil {
+				return fmt.Errorf("failed to set policy_statements on webhook with id %q: %v", webhookID, err)
+			}
+		} else {
+			err = d.Set(STATEMENTS, statements)
+			if err != nil {
+				return fmt.Errorf("failed to set statements on webhook with id %q: %v", webhookID, err)
+			}
+		}
 	}
 
 	err = d.Set(TAGS, webhook.Tags)